Implemented Token Validator

+++ src/main/java/com/example/productservice/Controllers/CategoryController.java
@@ -2,10 +2,14 @@ package com.example.productservice.Controllers;
 
 import com.example.productservice.DTOs.CategoryDTO;
 import com.example.productservice.Models.Category;
+import com.example.productservice.Security.JWTObject;
+import com.example.productservice.Security.TokenValidator;
 import com.example.productservice.Services.SelfCategoryService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.lang.Nullable;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.LinkedList;
@@ -21,11 +25,22 @@ public class CategoryController{
     @Autowired
     private SelfCategoryService selfCategoryService;
 
+    @Autowired
+    private TokenValidator tokenValidator;
+
     //Get All Available Category
     @GetMapping("")
-    public ResponseEntity<List<CategoryDTO>> getAllCategory(){
+    public ResponseEntity<List<CategoryDTO>> getAllCategory(@Nullable @RequestHeader(HttpHeaders.AUTHORIZATION) String token){
         ResponseEntity<List<CategoryDTO>> responseEntity;
         try{
+            JWTObject authtoken = null;
+            if(token !=  null) {
+                authtoken = tokenValidator.validateToken(token);
+            }
+            if(authtoken == null){
+//                System.out.println("Got token null");
+                return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+            }
             List<CategoryDTO> categoryDTOS = this.categorysToCategoryDTOs(selfCategoryService.getAllCategory());
             responseEntity = new ResponseEntity<>(categoryDTOS, HttpStatus.OK);
             return responseEntity;
@@ -37,9 +52,17 @@ public class CategoryController{
 
     //Get Category by id
     @GetMapping("/{id}")
-    public ResponseEntity<CategoryDTO> getSingleCategory(@PathVariable("id") Long categoryId){
+    public ResponseEntity<CategoryDTO> getSingleCategory(@Nullable @RequestHeader(HttpHeaders.AUTHORIZATION) String token, @PathVariable("id") Long categoryId){
         ResponseEntity<CategoryDTO> responseEntity;
         try{
+            JWTObject authtoken = null;
+            if(token !=  null) {
+                authtoken = tokenValidator.validateToken(token);
+            }
+            if(authtoken == null){
+//                System.out.println("Got token null");
+                return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+            }
             CategoryDTO categoryDTO = this.categoryDTOToCategoryDTO(selfCategoryService.getCategory(categoryId));
             responseEntity = new ResponseEntity<>(categoryDTO, HttpStatus.OK);
             return responseEntity;
@@ -51,9 +74,17 @@ public class CategoryController{
 
     //Add New Category
     @PostMapping(value = "")
-    public ResponseEntity<CategoryDTO> addNewCategory(@RequestBody CategoryDTO categoryDTORequest){
+    public ResponseEntity<CategoryDTO> addNewCategory(@Nullable @RequestHeader(HttpHeaders.AUTHORIZATION) String token, @RequestBody CategoryDTO categoryDTORequest){
         ResponseEntity<CategoryDTO> responseEntity;
         try{
+            JWTObject authtoken = null;
+            if(token !=  null) {
+                authtoken = tokenValidator.validateToken(token);
+            }
+            if(authtoken == null){
+//                System.out.println("Got token null");
+                return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+            }
             CategoryDTO categoryDTO = this.categoryDTOToCategoryDTO(selfCategoryService.addCategory(categoryDTORequest));
             responseEntity = new ResponseEntity<>(categoryDTO, HttpStatus.OK);
             return responseEntity;
@@ -65,9 +96,17 @@ public class CategoryController{
 
     //Delete Category
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteCategory(@PathVariable("id") Long categoryId){
+    public ResponseEntity<Void> deleteCategory(@Nullable @RequestHeader(HttpHeaders.AUTHORIZATION) String token, @PathVariable("id") Long categoryId){
         ResponseEntity<Void> responseEntity;
         try{
+            JWTObject authtoken = null;
+            if(token !=  null) {
+                authtoken = tokenValidator.validateToken(token);
+            }
+            if(authtoken == null){
+//                System.out.println("Got token null");
+                return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+            }
             selfCategoryService.deleteCategory(categoryId);
             return new ResponseEntity<>(HttpStatus.OK);
         }
@@ -78,9 +117,17 @@ public class CategoryController{
 
     //Update Category
     @PatchMapping("/{id}")
-    public ResponseEntity<CategoryDTO> patchProduct(@RequestBody CategoryDTO categoryDTORequest, @PathVariable("id") Long id){
+    public ResponseEntity<CategoryDTO> patchProduct(@Nullable @RequestHeader(HttpHeaders.AUTHORIZATION) String token, @RequestBody CategoryDTO categoryDTORequest, @PathVariable("id") Long id){
         ResponseEntity<CategoryDTO> responseEntity;
         try{
+            JWTObject authtoken = null;
+            if(token !=  null) {
+                authtoken = tokenValidator.validateToken(token);
+            }
+            if(authtoken == null){
+//                System.out.println("Got token null");
+                return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+            }
             CategoryDTO categoryDTO = this.categoryDTOToCategoryDTO(selfCategoryService.patchCategory(id, categoryDTORequest));
             responseEntity = new ResponseEntity<>(categoryDTO, HttpStatus.OK);
             return responseEntity;
